+++ include/exiv2/webpimage.hpp
@@ -40,7 +40,7 @@ class EXIV2API WebPImage : public Image {
   //@{
   void readMetadata() override;
   void writeMetadata() override;
-  void printStructure(std::ostream& out, PrintStructureOption option, int depth) override;
+  void printStructure(std::ostream& out, PrintStructureOption option, size_t depth) override;
   //@}
 
   /*!
