+++ src/webpimage.cpp
@@ -407,7 +407,7 @@ void WebPImage::doWriteMetadata(BasicIo& outIo) {
 
 /* =========================================== */
 
-void WebPImage::printStructure(std::ostream& out, PrintStructureOption option, int depth) {
+void WebPImage::printStructure(std::ostream& out, PrintStructureOption option, size_t depth) {
   if (io_->open() != 0) {
     throw Error(ErrorCode::kerDataSourceOpenFailed, io_->path(), strError());
   }
@@ -450,7 +450,7 @@ void WebPImage::printStructure(std::ostream& out, PrintStructureOption option, i
       if (equalsWebPTag(chunkId, WEBP_CHUNK_HEADER_EXIF) && option == kpsRecursive) {
         // create memio object with the payload, then print the structure
         MemIo p(payload.c_data(), payload.size());
-        printTiffStructure(p, out, option, depth+1);
+        printTiffStructure(p, out, option, depth + 1);
       }
 
       bool bPrintPayload = (equalsWebPTag(chunkId, WEBP_CHUNK_HEADER_XMP) && option == kpsXMP) ||
